requirements.txt
django
django-crispy
django-vote